+++ ios/nodes/GainNode/IOSGainNode.h
@@ -13,8 +13,8 @@ namespace audiocontext {
     class IOSGainNode : public IOSAudioNode {
         public:
             explicit IOSGainNode(std::shared_ptr<IOSAudioContext> context);
-            void changeGain(const double gain) const;
-            double getGain() const;
+            void changeGain(const float gain) const;
+            float getGain() const;
 
         protected:
             GainNode *gainNode_;
